Adding in more comments as well as new support for additional exception.
Refactored the error handling a little when making API requests to Dotmailer.

Also added more detailed information to the various exception classes so that when they are raised by the connection object there is some more details as to the problem with the request for the user to debug against.  The information for these messages are taken directly from DotMailer's API page.